refactor: extracted code inot facilty and owner share actions, then made placeholder file for the future state profile
