+++ src/lib/shareability/stateShareActions.js
@@ -1,0 +1,8 @@
+/**
+ * stateShareActions
+ *
+ * State-profile export logic (placeholder). Will layer on the generic
+ * shareActions.js primitives and the shared helpers in profileShareActions.js
+ * (copy-link / generic CSV category), owning the state-specific data shaping —
+ * mirroring facilityShareActions.js and ownerShareActions.js. Nothing here yet.
+ */
